Finish logic, fixing tests

+++ src/main/java/org/opensearch/tasks/service/TasksService.java
@@ -14,7 +14,6 @@ import org.opensearch.core.rest.RestStatus;
 import org.opensearch.tasks.model.Tasks;
 import org.opensearch.tasks.repository.TasksRepository;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -22,6 +21,7 @@ import java.util.stream.Collectors;
 public class TasksService {
     private final TasksRepository tasksRepository;
     private static final Logger log = LogManager.getLogger(TasksService.class);
+
     public TasksService(TasksRepository tasksRepository) {
         this.tasksRepository = tasksRepository;
     }
@@ -29,7 +29,7 @@ public class TasksService {
     public Tasks createTask(Tasks tasks) {
         log.info("Creating task {}", tasks);
         IndexResponse result = tasksRepository.createTask(tasks);
-        if(result.status() == RestStatus.CREATED){
+        if (result.status() == RestStatus.CREATED) {
             tasks.setId(result.getId());
             log.info("Task creating result {}", tasks);
             return tasks;
@@ -39,22 +39,58 @@ public class TasksService {
     }
 
     public Tasks getTaskById(String id) {
-        return tasksRepository.getTaskById(id);
+        log.info("Getting task by id {}", id);
+        if (id == null) {
+            log.info("Id was null, returning null");
+            return null;
+        }
+        Tasks tasks = tasksRepository.getTaskById(id);
+        log.info("Task retrieved {}", tasks);
+        return tasks;
     }
 
     public RestStatus deleteTask(String id) {
-        //check previously if exists
-        return tasksRepository.deleteTask(id);
+        log.info("Deleting task by id {}", id);
+        if (id == null) {
+            log.info("Id was null, returning NOT_FOUND");
+            return RestStatus.NOT_FOUND;
+        }
+        log.info("Searching task to delete");
+        Tasks task = tasksRepository.getTaskById(id);
+        if (task != null) {
+            log.info("Task found, deleting {}", task);
+            RestStatus status = tasksRepository.deleteTask(id);
+            log.info("Delete task process with status {}", status);
+            return status;
+        }
+        log.info("Task to delete not found");
+        return RestStatus.NOT_FOUND;
     }
 
     public List<Tasks> searchTasks(Map<String, Object> body) {
-        log.info("---------Searching tasks------------");
+        log.info("Searching tasks");
         List<Tasks> tasksList = tasksRepository.searchTasks(body);
-        log.info("---------Tasks found {}------------",tasksList);
-        log.info("---------Filtering tasks found with params------------");
+        log.info("Tasks found {}", tasksList);
+        log.info("Filtering tasks found with search parameters");
         List<Tasks> tasksResponse = filterTasksByContains(body, tasksList);
-        log.info("---------Tasks filtered {}------------",tasksResponse);
+        log.info("Tasks filtered {}", tasksResponse);
         return tasksResponse;
+    }
+
+    public RestStatus updateTask(Tasks task) {
+        log.info("Updating task {}", task);
+        if (task.getId() == null) {
+            log.info("Id cannot be null to update {}", task.getId());
+            return null;
+        }
+        Tasks taskToUpdate = tasksRepository.getTaskById(task.getId());
+        if (taskToUpdate == null) {
+            log.info("Task not found to update {}", task.getId());
+            return null;
+        }
+        IndexResponse taskUpdated = tasksRepository.updateTask(task);
+        log.info("Update task process with status {}", taskUpdated);
+        return taskUpdated.status();
     }
 
     private static List<Tasks> filterTasksByContains(Map<String, Object> body, List<Tasks> tasksList) {
@@ -67,28 +103,12 @@ public class TasksService {
         return tasksList.stream()
                 .filter(task ->
                         (!contains.containsKey("title") || task.getTitle().contains(contains.get("title").toString())) &&
-                        (!contains.containsKey("description") || task.getDescription().contains(contains.get("description").toString())) &&
-                        (!contains.containsKey("status") || task.getStatus().contains(contains.get("status").toString())) &&
-                        (!contains.containsKey("assignee") || task.getAssignee().contains(contains.get("assignee").toString())) &&
-                        (tagsToFilter == null || task.getTags().stream().allMatch(tagsToFilter::contains))
-        ).collect(Collectors.toList());
+                                (!contains.containsKey("description") || task.getDescription().contains(contains.get("description").toString())) &&
+                                (!contains.containsKey("status") || task.getStatus().contains(contains.get("status").toString())) &&
+                                (!contains.containsKey("assignee") || task.getAssignee().contains(contains.get("assignee").toString())) &&
+                                (tagsToFilter == null || task.getTags().stream().allMatch(tagsToFilter::contains))
+                ).collect(Collectors.toList());
     }
 
-    public static Map<String, List<String>> filterByTags(Map<String, List<String>> tags, List<String> tagsToFilter) {
-        Map<String, List<String>> filteredMap = new HashMap<>();
 
-        for (Map.Entry<String, List<String>> entry : tags.entrySet()) {
-            List<String> tagList = entry.getValue();
-            if (tagList.containsAll(tagsToFilter)) {
-                filteredMap.put(entry.getKey(), tagList);
-            }
-        }
-
-        return filteredMap;
-    }
-
-    public RestStatus updateTask(Tasks task) {
-        //check previously if exists
-        tasksRepository.updateTask(task); return RestStatus.OK;
-    }
 }
